requirements.txt
discord.py==2.5.2
googletrans==4.0.0rc1
beautifulsoup4
flask
requests
newspaper3k
lxml_html_clean
python-dotenv>=1.0.0
pytest>=7.0.0
PyNaCl>=1.5.0
boto3>=1.26.0
